split the sklearn module
