+++ tests/sklearn/test_ScikitLearnHistTabularTrees.py
@@ -1,7 +1,9 @@
 import pandas as pd
 
 from tabular_trees import BaseModelTabularTrees
-from tabular_trees.sklearn import ScikitLearnHistTabularTrees
+from tabular_trees.sklearn.scikit_learn_hist_tabular_trees import (
+    ScikitLearnHistTabularTrees,
+)
 
 
 def test_inheritance():
