+++ tests/sklearn/test_export_tree_data__hist_gradient_boosting_model.py
@@ -3,7 +3,7 @@ import re
 import pandas as pd
 import pytest
 
-from tabular_trees import sklearn
+from tabular_trees.sklearn import scikit_learn_hist_tabular_trees
 
 
 def test_model_not_correct_type_exception():
@@ -18,7 +18,9 @@ def test_model_not_correct_type_exception():
     )
 
     with pytest.raises(TypeError, match=re.escape(msg)):
-        sklearn._export_tree_data__hist_gradient_boosting_model(["a", "b"])
+        scikit_learn_hist_tabular_trees._export_tree_data__hist_gradient_boosting_model(
+            ["a", "b"]
+        )
 
 
 def test_multiple_responses_exception(sklearn_iris_hist_gbm_classifier):
@@ -30,7 +32,7 @@ def test_multiple_responses_exception(sklearn_iris_hist_gbm_classifier):
     with pytest.raises(
         NotImplementedError, match="model with multiple responses not supported"
     ):
-        sklearn._export_tree_data__hist_gradient_boosting_model(
+        scikit_learn_hist_tabular_trees._export_tree_data__hist_gradient_boosting_model(
             sklearn_iris_hist_gbm_classifier
         )
 
@@ -59,15 +61,20 @@ def test_output(mocker, sklearn_diabetes_hist_gbm_regressor):
     )
 
     mocker.patch.object(
-        sklearn, "_extract_hist_gbm_tree_data", return_value=dummy_tree_data
+        scikit_learn_hist_tabular_trees,
+        "_extract_hist_gbm_tree_data",
+        return_value=dummy_tree_data,
     )
 
-    exported_trees = sklearn._export_tree_data__hist_gradient_boosting_model(
-        sklearn_diabetes_hist_gbm_regressor
+    exported_trees = (
+        scikit_learn_hist_tabular_trees._export_tree_data__hist_gradient_boosting_model(
+            sklearn_diabetes_hist_gbm_regressor
+        )
     )
 
     assert (
-        type(exported_trees) is sklearn.ScikitLearnHistTabularTrees
+        type(exported_trees)
+        is scikit_learn_hist_tabular_trees.ScikitLearnHistTabularTrees
     ), "output from _export_tree_data__hist_gradient_boosting_model incorrect type"
 
     pd.testing.assert_frame_equal(exported_trees.trees, dummy_tree_data)
